Added notification system

+++ app/Http/Controllers/FriendController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\User;
 use App\Friend;
 use App\FriendRequest;
+use App\Notification;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -18,8 +19,18 @@ class FriendController extends Controller
         $friendRequest->uid1 = $uid1;
         $friendRequest->uid2 = $uid2;
 
+        // Notification setup
+        $notif = new Notification();
+        $notif->user_id = Auth::user()->id;
+        $notif->to = $uid2;
+        $notif->type = 'fRequest';
+
         if($friendRequest->save()){
-            return redirect()->back()->with(['message' => 'Friend Request successfully send.']);
+            if ($notif->save()) {
+                return redirect()->back()->with(['message' => 'Friend Request successfully send.']);
+            }else{
+                return redirect()->back()->with(['message' => 'There was an error. Please try again later']);
+            }
         }else{
             return redirect()->back()->with(['message' => 'Could not send friend request. Please try again later']);
         }
@@ -46,9 +57,19 @@ class FriendController extends Controller
         $friend->uid1 = $friendRequest->uid1;
         $friend->uid2 = $friendRequest->uid2;
 
+        // Notification setup
+        $notif = new Notification();
+        $notif->user_id = $friendRequest->uid2;
+        $notif->to = $friendRequest->uid1;
+        $notif->type = 'fRequestAccept';
+
         if($friend->save()){
-            $friendRequest->delete();
-            return redirect()->route('dashboard');
+            if($notif->save()) {
+                $friendRequest->delete();
+                return redirect()->route('dashboard');
+            }else{
+                return redirect()->route('dashboard')->with(['message' => 'There was a problem sending the notification.']);
+            }
         }else{
             $message = "Could not accept request. Please try again later";
             return redirect()->back()->with(['message' => $message]);
